TASK: Create UsageController in the Media Browser
This change is first iteration to refactor the AssetController and
make it slimer and easier to work with. The current AssetController
as a lots of responsability. This first step extract the relatedNodes
action to a specific controller. We need to forward the request from
the AssetController, because of how we use the routing in backend module
