+++ Neos.Media.Browser/Classes/Controller/AssetController.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace Neos\Media\Browser\Controller;
 
 /*
@@ -13,13 +14,6 @@ namespace Neos\Media\Browser\Controller;
 
 use Doctrine\Common\Persistence\Proxy as DoctrineProxy;
 use Doctrine\ORM\EntityNotFoundException;
-use Neos\ContentRepository\Domain\Model\NodeInterface;
-use Neos\ContentRepository\Domain\Model\Workspace;
-use Neos\ContentRepository\Domain\Repository\NodeDataRepository;
-use Neos\ContentRepository\Domain\Repository\WorkspaceRepository;
-use Neos\ContentRepository\Domain\Service\NodeTypeManager;
-use Neos\ContentRepository\Exception\NodeTypeNotFoundException;
-use Neos\Eel\FlowQuery\FlowQuery;
 use Neos\Error\Messages\Error;
 use Neos\Error\Messages\Message;
 use Neos\Flow\Annotations as Flow;
@@ -40,8 +34,8 @@ use Neos\Media\Domain\Model\AssetCollection;
 use Neos\Media\Domain\Model\AssetInterface;
 use Neos\Media\Domain\Model\AssetSource\AssetNotFoundExceptionInterface;
 use Neos\Media\Domain\Model\AssetSource\AssetProxyRepositoryInterface;
-use Neos\Media\Domain\Model\AssetSource\AssetSourceInterface;
 use Neos\Media\Domain\Model\AssetSource\AssetSourceConnectionExceptionInterface;
+use Neos\Media\Domain\Model\AssetSource\AssetSourceInterface;
 use Neos\Media\Domain\Model\AssetSource\AssetTypeFilter;
 use Neos\Media\Domain\Model\AssetSource\Neos\NeosAssetProxy;
 use Neos\Media\Domain\Model\AssetSource\SupportsCollectionsInterface;
@@ -55,12 +49,8 @@ use Neos\Media\Domain\Service\AssetService;
 use Neos\Media\TypeConverter\AssetInterfaceConverter;
 use Neos\Neos\Controller\BackendUserTranslationTrait;
 use Neos\Neos\Controller\CreateContentContextTrait;
-use Neos\Neos\Domain\Model\Dto\AssetUsageInNodeProperties;
 use Neos\Neos\Domain\Repository\DomainRepository;
 use Neos\Neos\Domain\Repository\SiteRepository;
-use Neos\Neos\Domain\Service\ContentDimensionPresetSourceInterface;
-use Neos\Neos\Domain\Service\UserService as DomainUserService;
-use Neos\Neos\Service\UserService;
 use Neos\Utility\Files;
 use Neos\Utility\MediaTypes;
 
@@ -91,18 +81,6 @@ class AssetController extends ActionController
 
     /**
      * @Flow\Inject
-     * @var NodeDataRepository
-     */
-    protected $nodeDataRepository;
-
-    /**
-     * @Flow\Inject
-     * @var NodeTypeManager
-     */
-    protected $nodeTypeManager;
-
-    /**
-     * @Flow\Inject
      * @var SiteRepository
      */
     protected $siteRepository;
@@ -112,30 +90,6 @@ class AssetController extends ActionController
      * @var DomainRepository
      */
     protected $domainRepository;
-
-    /**
-     * @Flow\Inject()
-     * @var WorkspaceRepository
-     */
-    protected $workspaceRepository;
-
-    /**
-     * @Flow\Inject
-     * @var UserService
-     */
-    protected $userService;
-
-    /**
-     * @Flow\Inject
-     * @var DomainUserService
-     */
-    protected $domainUserService;
-
-    /**
-     * @Flow\Inject
-     * @var ContentDimensionPresetSourceInterface
-     */
-    protected $contentDimensionPresetSource;
 
     /**
      * @Flow\Inject
@@ -394,7 +348,7 @@ class AssetController extends ActionController
             throw new \RuntimeException('Given asset source is not configured.', 1509632166);
         }
 
-        $assetSource =  $this->assetSources[$assetSourceIdentifier];
+        $assetSource = $this->assetSources[$assetSourceIdentifier];
         $assetProxyRepository = $assetSource->getAssetProxyRepository();
 
         try {
@@ -502,7 +456,7 @@ class AssetController extends ActionController
         $assetMappingConfiguration = $this->arguments->getArgument('asset')->getPropertyMappingConfiguration();
         $assetMappingConfiguration->allowProperties('title', 'resource');
         $assetMappingConfiguration->setTypeConverterOption(PersistentObjectConverter::class, PersistentObjectConverter::CONFIGURATION_CREATION_ALLOWED, true);
-        $assetMappingConfiguration->setTypeConverterOption(AssetInterfaceConverter::class, AssetInterfaceConverter::CONFIGURATION_ONE_PER_RESOURCE, true);
+        $assetMappingConfiguration->setTypeConverterOption(AssetInterfacseConverter::class, AssetInterfaceConverter::CONFIGURATION_ONE_PER_RESOURCE, true);
     }
 
     /**
@@ -728,95 +682,6 @@ class AssetController extends ActionController
     }
 
     /**
-     * Get Related Nodes for an asset
-     *
-     * @param AssetInterface $asset
-     * @return void
-     */
-    public function relatedNodesAction(AssetInterface $asset)
-    {
-        $userWorkspace = $this->userService->getPersonalWorkspace();
-
-        $usageReferences = $this->assetService->getUsageReferences($asset);
-        $relatedNodes = [];
-        $inaccessibleRelations = [];
-
-        $existingSites = $this->siteRepository->findAll();
-
-        foreach ($usageReferences as $usage) {
-            $inaccessibleRelation = [
-                'type' => get_class($usage),
-                'usage' => $usage
-            ];
-
-            if (!$usage instanceof AssetUsageInNodeProperties) {
-                $inaccessibleRelations[] = $inaccessibleRelation;
-                continue;
-            }
-
-            try {
-                $nodeType = $this->nodeTypeManager->getNodeType($usage->getNodeTypeName());
-            } catch (NodeTypeNotFoundException $e) {
-                $nodeType = null;
-            }
-            /** @var Workspace $workspace */
-            $workspace = $this->workspaceRepository->findByIdentifier($usage->getWorkspaceName());
-            $accessible = $this->domainUserService->currentUserCanReadWorkspace($workspace);
-
-            $inaccessibleRelation['nodeIdentifier'] = $usage->getNodeIdentifier();
-            $inaccessibleRelation['workspaceName'] = $usage->getWorkspaceName();
-            $inaccessibleRelation['workspace'] = $workspace;
-            $inaccessibleRelation['nodeType'] = $nodeType;
-            $inaccessibleRelation['accessible'] = $accessible;
-
-            if (!$accessible) {
-                $inaccessibleRelations[] = $inaccessibleRelation;
-                continue;
-            }
-
-            $node = $this->getNodeFrom($usage);
-            // this should actually never happen.
-            if (!$node) {
-                $inaccessibleRelations[] = $inaccessibleRelation;
-                continue;
-            }
-
-            $flowQuery = new FlowQuery([$node]);
-            $documentNode = $flowQuery->closest('[instanceof Neos.Neos:Document]')->get(0);
-            // this should actually never happen, too.
-            if (!$documentNode) {
-                $inaccessibleRelations[] = $inaccessibleRelation;
-                continue;
-            }
-
-            $site = $node->getContext()->getCurrentSite();
-            foreach ($existingSites as $existingSite) {
-                /** @var Site $existingSite **/
-                $siteNodePath = '/sites/' . $existingSite->getNodeName();
-                if ($siteNodePath === $node->getPAth() || strpos($node->getPath(), $siteNodePath . '/') === 0) {
-                    $site = $existingSite;
-                }
-            }
-
-            $relatedNodes[$site->getNodeName()]['site'] = $site;
-            $relatedNodes[$site->getNodeName()]['nodes'][] = [
-                'node' => $node,
-                'documentNode' => $documentNode
-            ];
-        }
-
-        $this->view->assignMultiple([
-            'totalUsageCount' => count($usageReferences),
-            'nodeUsageClass' => AssetUsageInNodeProperties::class,
-            'asset' => $asset,
-            'inaccessibleRelations' => $inaccessibleRelations,
-            'relatedNodes' => $relatedNodes,
-            'contentDimensions' => $this->contentDimensionPresetSource->getAllPresets(),
-            'userWorkspace' => $userWorkspace
-        ]);
-    }
-
-    /**
      * @param AssetCollection $assetCollection
      * @return void
      */
@@ -833,6 +698,17 @@ class AssetController extends ActionController
         $this->assetCollectionRepository->remove($assetCollection);
         $this->addFlashMessage('collectionHasBeenDeleted', '', Message::SEVERITY_OK, [htmlspecialchars($assetCollection->getTitle())]);
         $this->redirect('index');
+    }
+
+    /**
+     * Get Related Nodes for an asset (proxy action)
+     *
+     * @param AssetInterface $asset
+     * @return void
+     */
+    public function relatedNodesAction(AssetInterface $asset)
+    {
+        $this->forward('relatedNodes', 'Usage', 'Neos.Media.Browser', ['asset' => $asset]);
     }
 
     /**
@@ -898,16 +774,6 @@ class AssetController extends ActionController
     private function getMaximumFileUploadSize()
     {
         return min(Files::sizeStringToBytes(ini_get('post_max_size')), Files::sizeStringToBytes(ini_get('upload_max_filesize')));
-    }
-
-    /**
-     * @param AssetUsageInNodeProperties $assetUsage
-     * @return NodeInterface
-     */
-    private function getNodeFrom(AssetUsageInNodeProperties $assetUsage)
-    {
-        $context = $this->_contextFactory->create(['workspaceName' => $assetUsage->getWorkspaceName(), 'dimensions' => $assetUsage->getDimensionValues(), 'invisibleContentShown' => true, 'removedContentShown' => true]);
-        return $context->getNodeByIdentifier($assetUsage->getNodeIdentifier());
     }
 
     /**
@@ -1031,11 +897,11 @@ class AssetController extends ActionController
             switch ($this->browserState->get('sortBy')) {
                 case 'Name':
                     $assetProxyRepository->orderBy(['resource.filename' => $this->browserState->get('sortDirection')]);
-                break;
+                    break;
                 case 'Modified':
                 default:
                     $assetProxyRepository->orderBy(['lastModified' => $this->browserState->get('sortDirection')]);
-                break;
+                    break;
             }
         }
     }
