+++ Neos.Media.Browser/Configuration/Policy.yaml
@@ -6,6 +6,9 @@ privilegeTargets:
     'Neos.Media.Browser:ManageAssets':
       matcher: 'method(Neos\Media\Browser\Controller\(Asset|Image)Controller->(index|new|show|edit|update|initializeCreate|create|replaceAssetResource|updateAssetResource|initializeUpload|upload|tagAsset|delete|createTag|editTag|updateTag|deleteTag|addAssetToCollection|relatedNodes)Action())'
 
+    'Neos.Media.Browser:AssetUsage':
+      matcher: 'method(Neos\Media\Browser\Controller\UsageController->relatedNodesAction())'
+
     'Neos.Media.Browser:AssetProxyService':
       matcher: 'method(Neos\Media\Browser\Controller\AssetProxyController->(index|import)Action())'
 
@@ -20,6 +23,9 @@ roles:
     privileges:
       -
         privilegeTarget: 'Neos.Media.Browser:ManageAssets'
+        permission: GRANT
+      -
+        privilegeTarget: 'Neos.Media.Browser:AssetUsage'
         permission: GRANT
       -
         privilegeTarget: 'Neos.Media.Browser:AssetProxyService'
