+++ Neos.Media.Browser/Resources/Private/Templates/Asset/Edit.html
@@ -87,7 +87,9 @@
                         </tbody>
                     </table>
                     <f:if condition="{assetProxy.asset.inUse}">
-                        <f:link.action action="relatedNodes" arguments="{asset:assetProxy.asset}" class="neos-button">{neos:backend.translate(id: 'relatedNodes', quantity: '{assetProxy.asset.usageCount}', arguments: {0: assetProxy.asset.usageCount}, package: 'Neos.Media.Browser')}</f:link.action>
+                        <f:link.action controller="Usage" action="relatedNodes" arguments="{asset:assetProxy.asset}" class="neos-button">
+                          {neos:backend.translate(id: 'relatedNodes', quantity: '{assetProxy.asset.usageCount}', arguments: {0: assetProxy.asset.usageCount}, package: 'Neos.Media.Browser')}
+                        </f:link.action>
                     </f:if>
                 </fieldset>
             </div>
